// src/types/errors.ts
export enum Errors {
  MISSING_SEND_MESSAGES = "MISSING_SEND_MESSAGES",
  MISSING_MANAGE_ROLES = "MISSING_MANAGE_ROLES",
  MISSING_KICK_MEMBERS = "MISSING_KICK_MEMBERS",
  MISSING_VIEW_CHANNEL = "MISSING_VIEW_CHANNEL",
  MISSING_READ_MESSAGE_HISTORY = "MISSING_READ_MESSAGE_HISTORY",
  MISSING_MANAGE_NICKNAMES = "MISSING_MANAGE_NICKNAMES",
  MISSING_MUTE_MEMBERS = "MISSING_MUTE_MEMBERS",
  MISSING_DEAFEN_MEMBERS = "MISSING_DEAFEN_MEMBERS",
  MISSING_SEND_TTS_MESSAGE = "MISSING_SEND_TTS_MESSAGE",
  MISSING_MANAGE_MESSAGES = "MISSING_MANAGE_MESSAGES",
  MISSING_MANAGE_CHANNELS = "MISSING_MANAGE_CHANNELS",
  MISSING_CREATE_INSTANT_INVITE = "MISSING_CREATE_INSTANT_INVITE",
  MISSING_MANAGE_WEBHOOKS = "MISSING_MANAGE_WEBHOOKS",
  MISSING_MANAGE_EMOJIS = "MISSING_MANAGE_EMOJIS",
  MISSING_BAN_MEMBERS = "MISSING_BAN_MEMBERS",
  MISSING_MANAGE_GUILD = "MISSING_MANAGE_GUILD",
  MISSING_VIEW_AUDIT_LOG = "MISSING_VIEW_AUDIT_LOG",
  MISSING_EMBED_LINKS = "MISSING_EMBED_LINKS",
  DELETE_MESSAGES_MIN = "DELETE_MESSAGES_MIN",
  MESSAGE_MAX_LENGTH = "MESSAGE_MAX_LENGTH",
  NICKNAMES_MAX_LENGTH = "NICKNAMES_MAX_LENGTH",
  PRUNE_MIN_DAYS = "PRUNE_MIN_DAYS",
  RATE_LIMIT_RETRY_MAXED = "RATE_LIMIT_RETRY_MAXED",
  MISSING_INTENT_GUILD_MEMBERS = "MISSING_INTENT_GUILD_MEMBERS",
  REQUEST_CLIENT_ERROR = "REQUEST_CLIENT_ERROR",
  REQUEST_SERVER_ERROR = "REQUEST_SERVER_ERROR",
  REQUEST_UNKNOWN_ERROR = "REQUEST_UNKNOWN_ERROR",
  BOTS_HIGHEST_ROLE_TOO_LOW = "BOTS_HIGHEST_ROLE_TOO_LOW",
  CHANNEL_NOT_IN_GUILD = "CHANNEL_NOT_IN_GUILD",
  INVALID_WEBHOOK_NAME = "INVALID_WEBHOOK_NAME",
  INVALID_WEBHOOK_OPTIONS = "INVALID_WEBHOOK_OPTIONS",
}